+++ NEWS.md
@@ -7,6 +7,7 @@
 ## New functions
 
 * `scale_weights()` to rescale design weights for multilevel models.
+* `pca()` and `pca_rotate()` to create tidy summaries of principal component analyses or rotated loadings matrices from PCA.
 
 # sjstats 0.12.0